+++ include/rendering/Mesh.h
@@ -32,7 +32,7 @@ public:
 
     void draw(const Shader& shader);
 
-    void getVertices(std::vector<glm::dvec3>& vertices) const;
+    void getVertices(std::vector<glm::vec3>& vertices) const;
 
     void getIndices(std::vector<unsigned int>& indices) const;
 };
